Fix circular dependency error
The circular dependency error that was happening for MainMvpPresenter
was corrected by modifying the provideMainPresenter method
in the ActivityModule.

It had an input parameter MainMvpPresenter<MainMvpView> and returned a
MainMvpPresenter<MainMvpView> object, which is the same kind. This was
causing the circular dependency and a variety of errors were showing up
at compile-time because of this.