Prepare response and apply checks to it
Before initializing the ResultsActivity, checks need to be applied to
the response to ensure that it is not null and contains at least one
object. With that confirmed, the response can be passed on to the
ResultsActivity.

An Adapter and a BaseViewHolder were created to prepare the UI of the
ResultsActivity.